+++ example/src/screens/HomeScreen.tsx
@@ -423,6 +423,14 @@ export default function HomeScreen() {
               }}
             />
           </View>
+          <View style={styles.buttonContainer}>
+            <Button
+              title="Multibanco"
+              onPress={() => {
+                navigation.navigate('MultibancoPaymentScreen');
+              }}
+            />
+          </View>
         </>
       </Collapse>